DeepLearning training and evaluation

+++ ImageSignalProcessing/run_isp.py
@@ -8,7 +8,7 @@
 # Features:
 #   - Robust config path handling (relative to script)
 #   - Flexible batch processing with per-set config
-#   - Direct ffmpeg streaming for video output 
+#   - Direct ffmpeg streaming for video output
 #   - User-friendly toggles for output (video, PNG)
 #   - Auto-tuning for optimal multiprocessing
 #   - User-settable video framerate
@@ -24,22 +24,23 @@
 # Date: 2025-06-26
 #
 
-import os
-import cv2
-import numpy as np
-from pipeline import Pipeline
-from utils.yacs import Config
-import rawpy
 import glob
-from collections import OrderedDict
-from tqdm import tqdm
-import shutil
-import ffmpeg
-import time
 import multiprocessing
+import os
+import shutil
+import time
+from collections import OrderedDict
+
+import cv2
+import ffmpeg
+import numpy as np
+import rawpy
+from pipeline import Pipeline
+from tqdm import tqdm
+from utils.yacs import Config
 
 # Set the default config directory globally
-CONFIG_DIR = os.path.join(os.path.dirname(__file__), 'configs')
+CONFIG_DIR = os.path.join(os.path.dirname(__file__), "configs")
 
 # Global toggles for output
 SAVE_VIDEO = True
@@ -51,6 +52,7 @@ NUM_PROCESSES = 12  # Set to "Auto" to enable auto-tuning, or an integer for man
 # Global setting for video framerate (user must set this)
 VIDEO_FRAMERATE = 30
 
+
 def getWhiteBalanceGains(rawFile, gainMult):
     """
     Extract white balance gains from a raw file and return as integers.
@@ -61,10 +63,11 @@ def getWhiteBalanceGains(rawFile, gainMult):
         [redGain, greenGain, blueGain]
     """
     (redGain, greenGain, blueGain, offset) = rawFile.camera_whitebalance
-    redGain = np.multiply(redGain, gainMult).astype('u2')
-    greenGain = np.multiply(greenGain, gainMult).astype('u2')
-    blueGain = np.multiply(blueGain, gainMult).astype('u2')
+    redGain = np.multiply(redGain, gainMult).astype("u2")
+    greenGain = np.multiply(greenGain, gainMult).astype("u2")
+    blueGain = np.multiply(blueGain, gainMult).astype("u2")
     return [redGain, greenGain, blueGain]
+
 
 def load_bayer(raw_path):
     """
@@ -84,9 +87,10 @@ def load_bayer(raw_path):
         raw = rawpy.imread(raw_path)
         bayer = np.asarray(raw.raw_image_visible)
         [redGain, greenGain, blueGain] = getWhiteBalanceGains(raw, gainMult=4)
-        data['RGB'] = [redGain, greenGain, blueGain]
-    data['bayer'] = bayer
+        data["RGB"] = [redGain, greenGain, blueGain]
+    data["bayer"] = bayer
     return bayer
+
 
 def process_image_dict(image_dict, out_dir):
     """
@@ -96,31 +100,46 @@ def process_image_dict(image_dict, out_dir):
         out_dir: directory to save output videos, or 'Parent' to use parent of input directory
     """
     for cfg_name, image_list in tqdm(image_dict.items(), desc="Configs", unit="config"):
-        cfg_filename = cfg_name + '.yaml'
+        cfg_filename = cfg_name + ".yaml"
         cfg_path = os.path.join(CONFIG_DIR, cfg_filename)
         set_name = cfg_name  # e.g. 'your_configA'
         # If image_list is a single directory, expand to all PNGs in that directory
-        if isinstance(image_list, str) or (isinstance(image_list, list) and len(image_list) == 1 and os.path.isdir(image_list[0])):
+        if isinstance(image_list, str) or (
+            isinstance(image_list, list) and len(image_list) == 1 and os.path.isdir(image_list[0])
+        ):
             input_dir = image_list if isinstance(image_list, str) else image_list[0]
-            print(f"\n[DEBUG] Expanding directory for set '{set_name}':\n    Input directory: {input_dir}")
-            image_list = sorted(glob.glob(os.path.join(input_dir, '*.png')))
-        print(f"\n[DEBUG] Image list for set '{set_name}':\n    {len(image_list)} images\n    First 5: {image_list[:5] if image_list else '[]'}\n    ...\n")
+            print(
+                f"\n[DEBUG] Expanding directory for set '{set_name}':\n    Input directory: {input_dir}"
+            )
+            image_list = sorted(glob.glob(os.path.join(input_dir, "*.png")))
+        print(
+            f"\n[DEBUG] Image list for set '{set_name}':\n    {len(image_list)} images\n    First 5: {image_list[:5] if image_list else '[]'}\n    ...\n"
+        )
         # Determine output directory
-        if out_dir == 'Parent' and image_list:
+        if out_dir == "Parent" and image_list:
             # Use parent of first image's directory
             parent_dir = os.path.dirname(os.path.dirname(image_list[0]))
             this_out_dir = parent_dir
         else:
             this_out_dir = out_dir
         os.makedirs(this_out_dir, exist_ok=True)
-        print(f"[DEBUG] Output directory for set '{set_name}':\n    Config: {cfg_path}\n    Output PNGs: {os.path.abspath(os.path.join(this_out_dir, f'{set_name}-HV'))}\n    Output video: {os.path.abspath(os.path.join(this_out_dir, f'{set_name}.mp4')) if SAVE_VIDEO else '[SKIPPED]'}\n")
+        print(
+            f"[DEBUG] Output directory for set '{set_name}':\n    Config: {cfg_path}\n    Output PNGs: {os.path.abspath(os.path.join(this_out_dir, f'{set_name}-HV'))}\n    Output video: {os.path.abspath(os.path.join(this_out_dir, f'{set_name}.mp4')) if SAVE_VIDEO else '[SKIPPED]'}\n"
+        )
         pipeline = Pipeline(Config(cfg_path))
         save_dir = os.path.join(this_out_dir, f"{set_name}-HV")
         os.makedirs(save_dir, exist_ok=True)
-        pipeline.batch_run(image_list, save_dir, load_bayer, suffixes='', num_processes=NUM_PROCESSES)
+        pipeline.batch_run(
+            image_list, save_dir, load_bayer, suffixes="", num_processes=NUM_PROCESSES
+        )
         images_out = glob.glob(f"{save_dir}/*.png")
-        print(f"[DEBUG] Processed PNGs for set '{set_name}':\n    {len(images_out)} images\n    First 5: {images_out[:5] if images_out else '[]'}\n    ...\n")
-        img_info = [[os.path.basename(img), float(os.path.basename(img).split('-')[0])] for img in images_out]
+        print(
+            f"[DEBUG] Processed PNGs for set '{set_name}':\n    {len(images_out)} images\n    First 5: {images_out[:5] if images_out else '[]'}\n    ...\n"
+        )
+        img_info = [
+            [os.path.basename(img), float(os.path.basename(img).split("-")[0])]
+            for img in images_out
+        ]
         img_info.sort(key=lambda x: x[1])
         if not img_info:
             print(f"[DEBUG] No images found for set '{set_name}'. Skipping.\n")
@@ -131,20 +150,29 @@ def process_image_dict(image_dict, out_dir):
         h, w = first_img.shape[:2]
         if SAVE_VIDEO:
             out_video_path = os.path.join(this_out_dir, f"{set_name}.mp4")
-            print(f"[DEBUG] Starting video writing for set '{set_name}':\n    Output: {out_video_path}\n    Frame size: {w}x{h}\n    Framerate: {VIDEO_FRAMERATE}\n")
+            print(
+                f"[DEBUG] Starting video writing for set '{set_name}':\n    Output: {out_video_path}\n    Frame size: {w}x{h}\n    Framerate: {VIDEO_FRAMERATE}\n"
+            )
             process = (
-                ffmpeg
-                .input('pipe:', format='rawvideo', pix_fmt='bgr24', s=f'{w}x{h}', framerate=VIDEO_FRAMERATE)
-                .output(out_video_path, vcodec='libx264', crf=12, pix_fmt='yuv420p')
+                ffmpeg.input(
+                    "pipe:",
+                    format="rawvideo",
+                    pix_fmt="bgr24",
+                    s=f"{w}x{h}",
+                    framerate=VIDEO_FRAMERATE,
+                )
+                .output(out_video_path, vcodec="libx264", crf=12, pix_fmt="yuv420p")
                 .overwrite_output()
-                .global_args('-loglevel', 'error')  # Suppress ffmpeg output
+                .global_args("-loglevel", "error")  # Suppress ffmpeg output
                 .run_async(pipe_stdin=True)
             )
         for idx, img in enumerate(tqdm(img_info, desc=f"Frames ({set_name})", unit="frame")):
             frame_path = os.path.join(save_dir, img[0])
             frame = cv2.imread(frame_path)
             if frame.shape[:2] != (h, w):
-                print(f"[DEBUG]    Resizing frame {os.path.basename(frame_path)}: {frame.shape[:2]} -> ({h}, {w})")
+                print(
+                    f"[DEBUG]    Resizing frame {os.path.basename(frame_path)}: {frame.shape[:2]} -> ({h}, {w})"
+                )
                 frame = cv2.resize(frame, (w, h))
             if SAVE_VIDEO:
                 process.stdin.write(frame.astype(np.uint8).tobytes())
@@ -152,11 +180,12 @@ def process_image_dict(image_dict, out_dir):
             process.stdin.close()
             process.wait()
             print(f"[DEBUG] Video saved for set '{set_name}': {out_video_path}\n")
-        
+
         if not SAVE_PNG:
             print(f"[DEBUG] Removing intermediate PNG directory for set '{set_name}': {save_dir}\n")
             shutil.rmtree(save_dir)
     print(f"\n[DEBUG] All videos saved to: {out_dir if out_dir != 'Parent' else this_out_dir}\n")
+
 
 def process_and_show_single_image(image_path, config_path, out_path=None):
     """
@@ -170,7 +199,7 @@ def process_and_show_single_image(image_path, config_path, out_path=None):
     temp_dir = "./_single_image_temp/"
     os.makedirs(temp_dir, exist_ok=True)
     # Use the same loader as batch, but for a single image
-    pipeline.batch_run([image_path], temp_dir, load_bayer, suffixes='', num_processes=1)
+    pipeline.batch_run([image_path], temp_dir, load_bayer, suffixes="", num_processes=1)
     # Find the output image
     out_imgs = glob.glob(f"{temp_dir}/*.png")
     if not out_imgs:
@@ -186,7 +215,10 @@ def process_and_show_single_image(image_path, config_path, out_path=None):
         print(f"Saved processed image to {out_path}")
     shutil.rmtree(temp_dir)
 
-def find_optimal_num_processes(raw_paths, save_dir, load_raw_fn, pipeline_class, config, max_procs=None, test_count=18):
+
+def find_optimal_num_processes(
+    raw_paths, save_dir, load_raw_fn, pipeline_class, config, max_procs=None, test_count=18
+):
     """
     Benchmark different process counts and return the optimal number for batch_run.
     Args:
@@ -203,7 +235,7 @@ def find_optimal_num_processes(raw_paths, save_dir, load_raw_fn, pipeline_class,
     if max_procs is None:
         max_procs = multiprocessing.cpu_count()
     test_paths = raw_paths[:test_count]
-    best_time = float('inf')
+    best_time = float("inf")
     best_proc = 1
     print("[BENCH] Benchmarking process counts...")
     for n_proc in [4, 6, 8, max_procs]:
@@ -212,7 +244,9 @@ def find_optimal_num_processes(raw_paths, save_dir, load_raw_fn, pipeline_class,
             continue
         pipeline = pipeline_class(config)
         start = time.time()
-        pipeline.batch_run(test_paths, save_dir, load_raw_fn, num_processes=n_proc, show_progress=True)
+        pipeline.batch_run(
+            test_paths, save_dir, load_raw_fn, num_processes=n_proc, show_progress=True
+        )
         elapsed = time.time() - start
         print(f"[BENCH] {n_proc} processes: {elapsed:.2f}s")
         if elapsed < best_time:
@@ -220,6 +254,7 @@ def find_optimal_num_processes(raw_paths, save_dir, load_raw_fn, pipeline_class,
             best_proc = n_proc
     print(f"[BENCH] Optimal number of processes: {best_proc}")
     return best_proc
+
 
 def auto_set_num_processes(image_dict, config_dir, load_raw_fn, pipeline_class):
     """
@@ -229,25 +264,28 @@ def auto_set_num_processes(image_dict, config_dir, load_raw_fn, pipeline_class):
     global NUM_PROCESSES
     first_key = next(iter(image_dict))
     first_list = image_dict[first_key]
-    if isinstance(first_list, str) or (isinstance(first_list, list) and len(first_list) == 1 and os.path.isdir(first_list[0])):
+    if isinstance(first_list, str) or (
+        isinstance(first_list, list) and len(first_list) == 1 and os.path.isdir(first_list[0])
+    ):
         input_dir = first_list if isinstance(first_list, str) else first_list[0]
-        sample_images = sorted(glob.glob(os.path.join(input_dir, '*.png')))[:18]
+        sample_images = sorted(glob.glob(os.path.join(input_dir, "*.png")))[:18]
     else:
         sample_images = first_list[:20]
-    temp_bench_dir = './_bench_temp/'
+    temp_bench_dir = "./_bench_temp/"
     os.makedirs(temp_bench_dir, exist_ok=True)
     NUM_PROCESSES = find_optimal_num_processes(
         sample_images,
         temp_bench_dir,
         load_raw_fn,
         pipeline_class,
-        Config(os.path.join(config_dir, first_key + '.yaml')),
-        test_count=20
+        Config(os.path.join(config_dir, first_key + ".yaml")),
+        test_count=20,
     )
     shutil.rmtree(temp_bench_dir)
     print(f"[INFO] Using NUM_PROCESSES = {NUM_PROCESSES} for batch processing.\n")
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     # ----------------------
     # Select your mode below
     # ----------------------
@@ -256,9 +294,9 @@ if __name__ == '__main__':
 
     # 2. Batch process multiple sets of images using a dict:
     image_dict = {
-        'FLIR8.9': ['H:/DanganDataset-Formatted2/Day/DistanceTest/1/FLIR-8.9/PNG'],
+        "FLIR8.9": ["H:/DanganDataset-Formatted2/Day/DistanceTest/1/FLIR-8.9/PNG"],
     }
-    out_dir = 'Parent'
+    out_dir = "Parent"
 
     if NUM_PROCESSES == "Auto":
         auto_set_num_processes(image_dict, CONFIG_DIR, load_bayer, Pipeline)
